Comprises all commits to allow to define 3rd party cookie domain explicitly in piwik configuration Extend Cookie constructor to accept domain
In order to be able to set a 3rd party cookie's domain explicitly (so
that the browser accepts it and sets it to the correct domain), the
constructor was adapted to accept a cookie domain parameter. The
concerning usages of that constructor were adapted as well.
Now it is possible to put a line 'cookie_domain = ",mydomain.tld" ' into
the config.ini.php that will be used by piwik.

Cookie domain parsing adjusted

The changes make sure that cookie wildcard domains always follow the
pattern ".<domain_name>", independent of the domain that was configured
in the config.ini.php for the key "cookie_domain".

Add cookie_domain entry to global configuration